pyproject.toml
[project]
name = "learn-python"
version = "0.1.0"
description = "Learn Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "greenlet>=3.3.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]

# ============================================
# Ruff 配置 - 代码风格检查
# ============================================
[tool.ruff]
line-length = 88                  # 每行最大字符数（PEP 8 推荐 79，Black 默认 88）
target-version = "py312"          # 目标 Python 版本，用于检查语法兼容性

# 启用的规则类别
[tool.ruff.lint]
select = [
    "E",   # pycodestyle 错误（如格式问题）
    "F",   # pyflakes 错误（如未使用的变量）
    "I",   # isort - 自动整理 import 顺序
    "N",   # pep8-naming - 命名规范检查
    "W",   # pycodestyle 警告
    "UP",  # pyupgrade - 自动升级旧语法到新语法
    "B",   # flake8-bugbear - 常见 bug 模式检测
]

# ============================================
# Pyright 配置 - 静态类型检查
# ============================================
[tool.pyright]
pythonVersion = "3.12"            # Python 版本
typeCheckingMode = "strict"       # 检查模式：off/basic/strict
include = ["src"]                 # 检查的目录
exclude = [
    "**/__pycache__",             # 排除 Python 缓存目录
    ".venv",                      # 排除虚拟环境
    "**/tests",                   # 排除测试文件（可选）
]

# ============================================
# pytest 配置 - 测试框架
# ============================================
[tool.pytest.ini_options]
asyncio_mode = "auto"             # 自动识别异步测试
testpaths = ["tests"]             # 测试文件所在目录
pythonpath = ["src"]              # 添加 src 到 Python 路径（关键！）
addopts = "-v --tb=short"         # 默认参数：
                                  #   -v = 详细输出
                                  #   --tb=short = 简短错误追踪
# 其他常用选项（可选）
# --cov=src                      # 覆盖率统计
# --cov-report=html              # 生成 HTML 覆盖率报告
# -x                             # 遇到失败立即停止
# --maxfail=3                    # 最多失败 3 次
